Split test to its file
