+++ web/tls_config_test.go
@@ -16,7 +16,6 @@
 package web
 
 import (
-	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
@@ -585,68 +584,4 @@ func TestUsers(t *testing.T) {
 	for _, testInputs := range testTables {
 		t.Run(testInputs.Name, testInputs.Test)
 	}
-}
-
-// TestBasicAuthCache validates that the cache is working by calling a password
-// protected endpoint multiple times.
-func TestBasicAuthCache(t *testing.T) {
-	server := &http.Server{
-		Addr: port,
-		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Write([]byte("Hello World!"))
-		}),
-	}
-
-	done := make(chan struct{})
-	t.Cleanup(func() {
-		if err := server.Shutdown(context.Background()); err != nil {
-			t.Fatal(err)
-		}
-		<-done
-	})
-
-	go func() {
-		ListenAndServe(server, "testdata/tls_config_users_noTLS.good.yml", testlogger)
-		close(done)
-	}()
-
-	login := func(username, password string, code int) {
-		client := &http.Client{}
-		req, err := http.NewRequest("GET", "http://localhost"+port, nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-		req.SetBasicAuth(username, password)
-		r, err := client.Do(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if r.StatusCode != code {
-			t.Fatalf("bad return code, expected %d, got %d", code, r.StatusCode)
-		}
-	}
-
-	// Initial logins, checking that it just works.
-	login("alice", "alice123", 200)
-	login("alice", "alice1234", 401)
-
-	var (
-		start = make(chan struct{})
-		wg    sync.WaitGroup
-	)
-	wg.Add(200)
-	for i := 0; i < 100; i++ {
-		go func() {
-			<-start
-			login("alice", "alice123", 200)
-			wg.Done()
-		}()
-		go func() {
-			<-start
-			login("alice", "alice1234", 401)
-			wg.Done()
-		}()
-	}
-	close(start)
-	wg.Wait()
 }
